cyen final project
All I did to this code was take the original and make a function for seeing if the guess the player inputs is in the word or not; it is at the very end.  It still needs work (like subprograms that were not implemented yet), but for places like these I have placed appropriate comments saying what could be done to complete the task needed at that point.  Please look over it and make sure it makes sense haha and if you see anything that could be different or have questions don't hesitate to ask me